Breaking SplineAdopterBase into its own file.  And adjusting
