+++ src/QMCWaveFunctions/BsplineFactory/SplineR2RAdoptor.h
@@ -17,6 +17,7 @@
 
 #include <OhmmsSoA/Container.h>
 #include <spline2/MultiBspline.hpp>
+#include "QMCWaveFunctions/BsplineFactory/SplineAdoptorBase.h"
 
 namespace qmcplusplus
 {
