+++ src/QMCWaveFunctions/EinsplineAdoptor.h
@@ -114,102 +114,6 @@ inline Tensor<T,D> outerProductSymm(const TinyVector<T,D>& v, const TinyVector<T
 //  eval_e2iphi(kPoints.size(),phase.data(),eikr.data());
 //}
 
-/** base class any SplineAdoptor
- *
- * This handles SC and twist and declare storage for einspline
- */
-template<typename ST, unsigned D>
-struct SplineAdoptorBase
-{
-#if (__cplusplus >= 201103L)
-  using PointType=TinyVector<ST,D>;
-  using SingleSplineType=UBspline_3d_d;
-  using DataType=ST; 
-#else
-  typedef TinyVector<ST,D> PointType;
-  typedef UBspline_3d_d SingleSplineType;
-  typedef ST DataType;
-#endif
-  ///true if the computed values are complex
-  bool is_complex;
-  ///true, if it has only one k point and Gamma
-  bool is_gamma_only;
-  ///true, if it has only one k point and Gamma
-  bool is_soa_ready;
-  ///Index of this adoptor, when multiple adoptors are used for NUMA or distributed cases
-  size_t MyIndex;
-  ///number of unique orbitals
-  size_t nunique_orbitals;
-  ///first index of the SPOs this Spline handles
-  size_t first_spo;
-  ///last index of the SPOs this Spline handles
-  size_t last_spo;
-  ///sign bits at the G/2 boundaries
-  TinyVector<int,D>          HalfG;
-  ///\f$GGt=G^t G \f$, transformation for tensor in LatticeUnit to CartesianUnit, e.g. Hessian
-  Tensor<ST,D>               GGt;
-  CrystalLattice<ST,D>       SuperLattice;
-  CrystalLattice<ST,D>       PrimLattice;
-  /// flags to unpack sin/cos
-  std::vector<bool>               MakeTwoCopies;
-  ///kpoints for each unique orbitals
-  std::vector<TinyVector<ST,D> >  kPoints;
-  ///remap band
-  aligned_vector<int> BandIndexMap;
-  /// band offsets
-  std::vector<int> offset_real, offset_cplx;
-  ///name of the adoptor
-  std::string AdoptorName;
-  ///keyword used to match hdf5
-  std::string KeyWord;
-
-  SplineAdoptorBase()
-    :is_complex(false),is_gamma_only(false), is_soa_ready(false),
-    MyIndex(0),nunique_orbitals(0),first_spo(0),last_spo(0)
-  { }
-
-#if (__cplusplus >= 201103L)
-  SplineAdoptorBase(const SplineAdoptorBase& rhs)=default;
-#endif
-
-  inline void init_base(int n)
-  {
-    nunique_orbitals=n;
-    GGt=dot(transpose(PrimLattice.G),PrimLattice.G);
-    kPoints.resize(n);
-    MakeTwoCopies.resize(n);
-  }
-
-  ///remap kpoints to group general kpoints & special kpoints
-  int remap_kpoints()
-  {
-    std::vector<TinyVector<ST,D> >  k_copy(kPoints);
-    const int nk=kPoints.size();
-    BandIndexMap.resize(nk);
-    int nCB=0;
-    //two pass
-    for(int i=0; i<nk; ++i)
-    {
-      if(MakeTwoCopies[i]) 
-      {
-        kPoints[nCB]=k_copy[i];
-        BandIndexMap[i]=nCB++;
-      }
-    }
-    int nRealBands=nCB;
-    for(int i=0; i<nk; ++i)
-    {
-      if(!MakeTwoCopies[i]) 
-      {
-        kPoints[nRealBands]=k_copy[i];
-        BandIndexMap[i]=nRealBands++;
-      }
-    }
-    return nCB; //return the number of complex bands
-  }
-};
-
-
 /** BsplineSet<SplineAdoptor>, a SPOSetBase
  * @tparam SplineAdoptor implements evaluation functions that matched the storage requirements.
  *
